requirements.txt
absl-py==0.9.0
addict==2.2.1
ase==3.19.1
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
Connectivity==0.0.0
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
dotmap==1.3.14
Eel==0.12.3
flowiz==2.3.1
future==0.18.2
fvcore==0.1.1.post200513
gevent==20.5.0
gevent-websocket==0.10.1
google-auth==1.14.3
google-auth-oauthlib==0.4.1
googledrivedownloader==0.4
greenlet==0.4.15
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.0
isodate==0.6.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
llvmlite==0.33.0
lpips==0.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mmcv==0.5.6
mmdet==0.5.6+10e5d7d
mock==4.0.2
moviepy==1.0.3
networkx==2.4
numba==0.50.1
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
plyfile==0.7.2
portalocker==1.7.0
proglog==0.1.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
rdflib==5.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2020.5.11
torch==1.5.0
torch-cluster==1.5.6
torch-geometric==1.6.0
torch-scatter==2.0.5
torch-sparse==0.6.6
torch-spline-conv==1.2.0
torchvision==0.6.0
tqdm==4.46.0
urllib3==1.25.9
visualpriors==0.3.5
Werkzeug==1.0.1
whichcraft==0.6.1
yacs==0.1.7
yapf==0.30.0
zipp==3.1.0


